fix: restore transitive side-effect registrations
Keep pure modules side-effect-free by restoring compatibility imports from public wrappers, following the precedent established in #18717.
