+++ packages/dev/core/src/Bones/skeleton.functions.ts
@@ -3,7 +3,7 @@ import { Matrix } from "../Maths/math.vector.pure";
 import { type Scene } from "../scene";
 import { type TransformNode } from "core/Meshes/transformNode";
 import { Mesh } from "core/Meshes/mesh.pure";
-import { MeshBuilder } from "core/Meshes/meshBuilder";
+import { MeshBuilder } from "core/Meshes/meshBuilder.pure";
 import { VertexBuffer } from "core/Buffers/buffer.pure";
 import { Skeleton } from "./skeleton";
